requirements.txt
selenium
pytest
webdriver-manager
openpyxl
allure-pytest
